# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zerizeha API
  version: "1.0"
  description: Zerizeha backend API
servers:
  - url: /
paths:
  /api/ws/webrtc/{connectionId}:
    get:
      operationId: WebRTCWebSocket
      summary: webrtc signaling websocket
      description: WebSocket upgrade endpoint for Janus signaling (offer/answer/trickle)
      tags:
        - Voice
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: connectionId
          required: true
          schema:
            type: string
      responses:
        '101':
          description: Switching Protocols
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMap'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    ErrorMap:
      type: object
      additionalProperties:
        type: string
...